Move som package into gen package to avoid clash with PySOM
