+++ gen/method_gen.py
@@ -1,5 +1,5 @@
-from som.ast.basic import MsgSend, Literal
-from som.method import Method
+from gen.som.ast.basic import MsgSend, Literal
+from gen.som.method import Method
 
 
 def create_method_print_string(method_name, clazz, string):
