+++ gen/som/ast/basic.py
@@ -1,6 +1,6 @@
 from gen.generator import IND
-from som.ast.priority import Priority
-from som.util import combine_pattern_with_args
+from gen.som.ast.priority import Priority
+from gen.som.util import combine_pattern_with_args
 
 
 class Expression:
