+++ gen/som/util.py
